Modularize line chart
